REAL bullet fix, better debug_draw() support and lots of method call reordering (and more...)
entity.py: Made a separate method debug_draw() so it can be easily extended/overwritten so setting self.config.debug = True will enable all implemented debug_draw() methods, not just the basic one. Because of that I had to slightly modify draw() and draw_debug() methods of other classes.

bullet.py: Found the root of all evil regarding bullets... I was drawing the bullet on screen first and only after that updating it's position. Would be fine if I did it in this order everywhere, but I only did it here, so of course bullet position was delayed. Had to make a few changes in other files because of this.

gun.py: Added should_fire (bool) attribute, and modified the class, so now if the player/agent tries to shoot, it doesn't attempt to fire immediately but simply sets should_fire to true. Once the gun.tick() method is called it checks if should_fire is true, and if it is, it actually attemps to fire. So no matter when the player fires, the bullet will be fired only when the gun is updated (later that very same frame (considering handling is done before updating game state)).

cloudskimmer.py: tick(&draw) is now called after updating member info.

inventory.py: The cooldown overlay is now correctly drawn - the first frame overlay covers the entire slot, not one "frame"/time step too little (before we were decrementing cooldown before drawing the overlay (in gun.py), now we update it after (in inventory.py)).

item.py: Tried adding tick() method which would decrement its remaining_cooldown but because we need to update player state before updating inventory (in flappybird.py, as otherwise gun position is one frame behind, and the player is drawn over the gun), we would have the same problem (in inventory.py) as before - remaining_cooldown being decremented too soon. That's why I added decrement_cooldown() method, which needs to be called in a subclass, so we can decide exactly when to decrement. In inventory.py we decrement after drawing the overlay.

enemy_cloudskimmer_env.py: added get_bullet_positions() and some other changes