REMOVE PACMAN
Completely removed all Pacman related code. The Pacman "minigame" does not fit in at all. It's too random. It's even past the "so random it's funny" line. Maybe the very first time it's funny, but it gets very repetitive and boring really quickly. And the fact that it takes longer to complete a Pacman minigame than what the average length of Flappy Bird Plus game is, doesn't make sense.