refactor: simplify get_action_masks() method
Now we pass the entity (instead of observation_instance) and environment to the get_action_masks() method, as it's enough and simplifies things.

At first I removed the get_action_masks() method from Controllers entirely and moved the logic to the training Environments and called the method in the BaseModelController from the environment. That worked perfectly fine, like it does now, but it did not feel right at all. Controller should contain all methods that it needs for the agent to control the entity and get_action_masks() is one of them. If we wanted action masks just during training, then keeping the logic in the training Environment would be perfectly fine, as we wouldn't need action masks in the controllers at all, but because we want action masks during running the trained model as well, I think it's better to have the get_action_masks() method in controllers. For running the trained models, I think we should be able to modify the Controllers, so it doesn't create brand new environments of training environments for each new instance of a controller, but instead we pass the main FlappyBird environment (wrapped in GymEnv and normalized). I think that would use slighly less RAM for each additional instance of a controller (and make more sense logically) and then we wouldn't be able to directly call the get_action_masks() method of the training environment, cuz it wouldn't be initialized at all. Not a huge improvement, I don't even know if it would work. Just thinking... useless long ass commit message I probably won't ever see again (-_-)