CAN YOU USE THE SEEDS??
Tried only using Config.seed for all seed related logic, but figured out I need another variable to handle all seed logic "nicely", so I added handle_seed. Why? Well if we didn't have handle_seed, we could use seed = -1 (instead of current handle_seed=False), and it works everywhere, except when we want to start the training with a custom seed AND let SB3 handle incremental seeds for all spawned environments. If we put -1, it used -1 as model seed, which is invalid. If we put actual seed, like 42, it used that for the model, but when creating environments, it overrid the seed that SB3 set, to seed 42 for all environments, so they were exactly the same, instead of different.